lessonmodule lesson lesson file

+++ src/core/error/validation.filter.ts
@@ -21,7 +21,7 @@ export class MulterValidationExceptionFilter implements ExceptionFilter {
           })
         })
       } 
-      if(file['banner']){
+      if(file['banner'] && Array.isArray(file['banner']) && file['banner'][0] && file['banner'][0].path){
         const path = file['banner'][0].path
         fs.unlink(path,(err) =>{
           if(err){
@@ -29,21 +29,21 @@ export class MulterValidationExceptionFilter implements ExceptionFilter {
           }
         })
       }
-      if(file['introVideo']){
+      if(file['introVideo'] && Array.isArray(file['introVideo']) && file['introVideo'][0] && file['introVideo'][0].path){
         fs.unlink(file['introVideo'][0].path,(err) => {
           if(err) {
             console.log(err)
           }
         })
       }
-      if(file.path){
+      if(file && file.path){
         fs.unlink(file.path,(err) => {
           if(err) {
             console.log(err)
           }
         })
       }
-      if(file['video']){
+      if(file['video'] && file['video'].path){
         fs.unlink(file['video'],(err) => {
           if(err){
             console.log(err)
@@ -51,7 +51,7 @@ export class MulterValidationExceptionFilter implements ExceptionFilter {
         })
       }
     } catch (error) {
-      console.log(error)
+      console.log("MulterValidationExceptionFilter catch  - > ", error)
     }
 
     response.status(status).json({
